requirements.txt
asgiref==3.2.7
CacheControl==0.12.6
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.6
django-filter==2.2.0
django-redis==4.11.0
djangorestframework==3.11.0
djangorestframework-api-key==2.0.0
drf-api-tracking==1.6.0
drf-firebase-auth==0.0.8
firebase-admin==4.2.0
google-api-core==1.17.0
google-api-python-client==1.8.2
google-auth==1.14.2
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-storage==1.28.1
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.28.1
httplib2==0.17.3
idna==2.9
Markdown==3.2.1
msgpack==1.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
redis==3.5.0
requests==2.23.0
rsa==4.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
